feat: sdk now java generic
